ml/requirements.txt
torch>=2.2.0
ultralytics>=8.2.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
opencv-python>=4.9.0
Pillow>=10.3.0
anthropic>=0.25.0
numpy>=1.26.0
scikit-learn>=1.4.0
